CNO configuration manager
This adds a new capability to the CNO to manage configuration changes.
It adds a flag to the reconciler object to be able to track if there
are configuration changes and if they are still in progress.

It also adds a new function to handle these configuration changes.
This function has access to the prev and new configuration, so
it can implement different workflows based on the configuration
changes.

Because these workflows may require multiple operations, and
can not be done in a single cycle, this function has access to the
reconciler variable to keep it enable until it finish the changes.

At this moment is only used to handle the dual-stack conversion
in OVN, but it can be generalized for other use cases.

Signed-off-by: Antonio Ojea <aojea@redhat.com>